requirements.txt
asgiref==3.8.1
black==24.4.0
click==8.1.7
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
sqlparse==0.5.0
